# docs/swagger.yaml
basePath: /
definitions:
...
  dto.AttendanceResponse:
    properties:
      not_found_cards:
        items:
          type: string
        type: array
      saved_events:
        type: integer
      session_id:
        type: integer
    type: object
  dto.AttendanceScanRequest:
    properties:
      card_uid:
        type: string
      scanned_at:
        type: string
    required:
    - scanned_at
    type: object
  dto.AttendanceSessionRequest:
    properties:
      data:
        type: string
      finished_at:
        type: string
      room:
        type: string
      scans:
        items:
          $ref: '#/definitions/dto.AttendanceScanRequest'
        type: array
      source:
        type: string
      started_at:
        type: string
    required:
    - data
    - finished_at
    - scans
    - started_at
    type: object
...
  dto.StudentAttendanceRecordResponse:
    properties:
      date:
        type: string
      finished_at:
        type: string
      group:
        type: string
      lesson_type:
        type: string
      pair:
        type: integer
      present:
        type: boolean
      room:
        type: string
      scanned_at:
        type: string
      session_id:
        type: integer
      source:
        type: string
      started_at:
        type: string
      subject:
        type: string
      teacher:
        type: string
    type: object
  dto.StudentAttendanceResponse:
    properties:
      records:
        items:
          $ref: '#/definitions/dto.StudentAttendanceRecordResponse'
        type: array
      status:
        type: string
      student_id:
        type: integer
      summary:
        $ref: '#/definitions/dto.StudentAttendanceSummary'
    type: object
  dto.StudentAttendanceSummary:
    properties:
      absent_count:
        type: integer
      attendance_percent:
        type: number
      present_count:
        type: integer
      total_sessions:
        type: integer
    type: object
...
  http.AttendanceCreateResponse:
    properties:
      data:
        $ref: '#/definitions/dto.AttendanceResponse'
      status:
        example: created
        type: string
    type: object
  http.ErrorResponse:
    properties:
      error:
        example: invalid request body
        type: string
      message:
        example: course query parameter is required
        type: string
      status:
        example: error
        type: string
    type: object
...
info:
  contact: {}
  description: API for schedule import/read, attendance processing and study plan
    management.
  title: CourseJob API
  version: "1.0"
paths:
...
  /api/v1/attendance/sessions:
    post:
      consumes:
      - application/json
      description: Accepts a session with scans and stores attendance events.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Attendance session payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AttendanceSessionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/http.AttendanceCreateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Create attendance session
      tags:
      - Attendance
...
  /api/v1/student/me/attendance:
    get:
      description: Returns attendance records and summary for the current authenticated
        student.
      parameters:
      - description: Bearer access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Start date in YYYY-MM-DD
        in: query
        name: from
        type: string
      - description: End date in YYYY-MM-DD
        in: query
        name: to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StudentAttendanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/http.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/http.ErrorResponse'
      summary: Get my attendance
      tags:
      - Students
...
schemes:
- https
swagger: "2.0"